refactor: type the state-byte constants long at the source
The byte-count constants feed only the signed state dimension, so
declaring them long removes every initializer cast; C# 14 constant
conversions keep the few ulong-context uses working unchanged.

+++ src/Nethermind/Nethermind.Core/Eip8037Constants.cs
@@ -10,10 +10,10 @@ public static class Eip8037Constants
     public const ulong PerAuthBaseRegularCost = 7_500;
 
     /// <summary> State bytes charged per authorization tuple for the state-gas dimension. </summary>
-    public const ulong StateBytesPerAuthBase = 23;
+    public const long StateBytesPerAuthBase = 23;
 
     public const ulong SystemCallBaseGasLimit = 30_000_000;
-    public const ulong SystemMaxSstoresPerCall = 16;
-    public const long SystemCallStateReservoir = (long)(GasCostOf.StateBytesPerStorageSet * GasCostOf.CostPerStateByte * SystemMaxSstoresPerCall);
+    public const long SystemMaxSstoresPerCall = 16;
+    public const long SystemCallStateReservoir = GasCostOf.StateBytesPerStorageSet * GasCostOf.CostPerStateByte * SystemMaxSstoresPerCall;
     public const ulong SystemCallGasLimit = SystemCallBaseGasLimit + SystemCallStateReservoir;
 }
